+++ src/components/ListItem.jsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import './ListItem.css';
 import { updateItem } from '../api/firebase';
 import { useMutation } from 'react-query';
+import { getNextPurchasedDate } from '../utils';
 
 export function ListItem({ item, listPath }) {
 	const { id, totalPurchases, name, dateLastPurchased, dateNextPurchased } =
@@ -29,11 +30,16 @@ export function ListItem({ item, listPath }) {
 	});
 
 	async function markAsPurchased() {
+		const nextPurchasedDate = getNextPurchasedDate({
+			dateLastPurchased,
+			dateNextPurchased,
+			totalPurchases,
+		});
+
 		await updateItem(listPath, {
 			itemId: id,
 			totalPurchases: totalPurchases,
-			dateLastPurchased: dateLastPurchased,
-			dateNextPurchased: dateNextPurchased,
+			dateNextPurchased: nextPurchasedDate,
 		});
 	}
